Rename UserIdService into IdService and move to util
