+++ .gitignore
@@ -1,2 +1,3 @@
 /.idea/
 *.iml
+/example_experts.json
