+++ backend/src/main/java/com/danielwindel/util/ids/IdService.java
@@ -1,11 +1,11 @@
-package com.danielwindel.users;
+package com.danielwindel.util.ids;
 
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 @Service
-public class UserIdService {
+public class IdService {
     public String generateId() {
         return UUID.randomUUID().toString();
     }
